+++ lib/mvvm_service.dart
@@ -1,11 +1,14 @@
 library;
 
-export 'components/service_id.dart';
 export 'components/service_mode.dart';
 export 'components/service_subscription.dart';
 export 'components/service.dart';
+export 'components/state_id.dart';
+export 'components/state_mode.dart';
+export 'components/state_subscription.dart';
 
 export 'extensions/service_build_context.dart';
 export 'extensions/service_when.dart';
 
 export 'widgets/service_scope.dart';
+export 'widgets/state_scope.dart';